doplneni GlobalExceptionMiddleware a dalsi zmeny

+++ TestConsoleApp/Program.cs
@@ -55,8 +55,8 @@ namespace TestConsoleApp
                 Thread.Sleep(500);
 
                 int c = 0;
-                short ok;
-                short ng;
+                int ok;
+                int ng;
 
                 while (c++ < 10)
                 {
